project-website/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=0.7.1=py37_0
astor=0.8.0=py37_0
beautifulsoup4=4.8.0=pypi_0
blas=1.0=mkl
ca-certificates=2019.5.15=1
certifi=2019.6.16=py37_1
chardet=3.0.4=pypi_0
click=7.0=pypi_0
cycler=0.10.0=py37_0
django=2.2.1=py37_0
flask=1.1.1=py_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=py37_0
grpcio=1.16.1=py37h351948d_1
gtts=2.0.4=pypi_0
gtts-token=1.1.3=pypi_0
h5py=2.8.0=py37hf7173ca_2
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
intel-openmp=2019.5=281
itsdangerous=1.1.0=py37_0
jinja2=2.10.1=py37_0
jpeg=9c=hfa6e2cd_1001
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37ha925a31_0
libblas=3.8.0=12_mkl
libcblas=3.8.0=12_mkl
libclang=8.0.1=he980bc4_1
liblapack=3.8.0=12_mkl
liblapacke=3.8.0=12_mkl
libmklml=2019.0.5=0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.8.0=h7bd577a_0
libtiff=4.0.10=hb898794_2
libwebp=1.0.2=hfa6e2cd_2
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.14=py37h14836fe_0
mkl_random=1.0.2=py37h343c172_0
numpy=1.16.5=py37h19fb1c0_0
numpy-base=1.16.5=py37hc3f5095_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1d=he774522_0
pillow=6.1.0=py37hdc69c19_0
pip=19.2.2=py37_0
protobuf=3.8.0=py37h33f27b4_0
py-opencv=3.4.2=py37hc319ecb_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
python=3.7.4=h5263a28_0
python-dateutil=2.8.0=py37_0
pytz=2019.2=py_0
pyyaml=5.1.2=py37he774522_0
qt=5.9.7=vc14h73c81de_0
requests=2.22.0=pypi_0
scipy=1.3.1=py37h29ff71c_0
setuptools=41.0.1=py37_0
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
soupsieve=1.9.3=pypi_0
sqlite=3.29.0=he774522_0
sqlparse=0.3.0=py_0
tensorboard=1.14.0=py37he3c9ec2_0
tensorflow=1.14.0=mkl_py37h7908ca0_0
tensorflow-base=1.14.0=mkl_py37ha978198_0
tensorflow-estimator=1.14.0=py_0
termcolor=1.1.0=py37_1
tk=8.6.8=hfa6e2cd_0
tornado=6.0.3=py37he774522_0
urllib3=1.25.5=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
werkzeug=0.15.5=py_0
wheel=0.33.4=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
